Don't use environment object in static views
Static views which only display data passed to them do not need to use
the @EnvironmentObject. Instead, we can pass the data through the
constructor. The only views which need access to the dyanmic data are
the ContentView (as it selects the dynamic data) and the
NavigationBarDetailView (since it sets the data to nil which the back
button is tapped). All the other views only need to be provided static
data to be displayed.

This is an extension to the course which is aimed at simplifying the
application by:
 1. Not using the Ennvironment variable to access data in static views
    (as discussed above).
 2. Simplifying the testing for an invalid product, since this is only
    done once when passing to the ProductDetailView from the
    ContentView.
Note: It would be preferable to use a placeholder product rather than
the sampleProduct used in this application for when an invalid product
is passed.